Create a shared component for saga audits
Ale extracts some RavenDB-related code to a shared infra lib
