+++ src/ServiceControl.Audit.AcceptanceTests/SagaAudit/When_a_saga_instance_is_being_created.cs
@@ -4,10 +4,10 @@
     using System.Linq;
     using System.Threading.Tasks;
     using AcceptanceTesting;
-    using Audit.SagaAudit;
     using NServiceBus;
     using NServiceBus.AcceptanceTesting;
     using NUnit.Framework;
+    using ServiceControl.SagaAudit;
     using TestSupport.EndpointTemplates;
 
     [RunOnAllTransports]
