+++ src/ServiceControl.Audit.UnitTests/ApprovalFiles/APIApprovals.PublicClr.approved.txt
@@ -273,72 +273,6 @@ namespace ServiceControl.Audit.Recoverability
 }
 namespace ServiceControl.Audit.SagaAudit
 {
-    public class InitiatingMessage
-    {
-        public InitiatingMessage() { }
-        public string Intent { get; set; }
-        public bool IsSagaTimeoutMessage { get; set; }
-        public string MessageId { get; set; }
-        public string MessageType { get; set; }
-        public string OriginatingEndpoint { get; set; }
-        public string OriginatingMachine { get; set; }
-        public System.DateTime TimeSent { get; set; }
-    }
-    public class ResultingMessage
-    {
-        public ResultingMessage() { }
-        public System.Nullable<System.DateTime> DeliverAt { get; set; }
-        public System.Nullable<System.TimeSpan> DeliveryDelay { get; set; }
-        public string Destination { get; set; }
-        public string Intent { get; set; }
-        public string MessageId { get; set; }
-        public string MessageType { get; set; }
-        public System.DateTime TimeSent { get; set; }
-    }
-    public class SagaDetailsIndex : Raven.Client.Indexes.AbstractMultiMapIndexCreationTask<ServiceControl.Audit.SagaAudit.SagaHistory>
-    {
-        public SagaDetailsIndex() { }
-    }
-    public class SagaHistory
-    {
-        public SagaHistory() { }
-        public System.Collections.Generic.List<ServiceControl.Audit.SagaAudit.SagaStateChange> Changes { get; set; }
-        public System.Guid Id { get; set; }
-        public System.Guid SagaId { get; set; }
-        public string SagaType { get; set; }
-    }
-    public class SagaSnapshot
-    {
-        public SagaSnapshot() { }
-        public string Endpoint { get; set; }
-        public System.DateTime FinishTime { get; set; }
-        public System.Guid Id { get; set; }
-        public ServiceControl.Audit.SagaAudit.InitiatingMessage InitiatingMessage { get; set; }
-        public System.Collections.Generic.List<ServiceControl.Audit.SagaAudit.ResultingMessage> OutgoingMessages { get; set; }
-        public System.DateTime ProcessedAt { get; set; }
-        public System.Guid SagaId { get; set; }
-        public string SagaType { get; set; }
-        public System.DateTime StartTime { get; set; }
-        public string StateAfterChange { get; set; }
-        public ServiceControl.Audit.SagaAudit.SagaStateChangeStatus Status { get; set; }
-    }
-    public class SagaStateChange
-    {
-        public SagaStateChange() { }
-        public string Endpoint { get; set; }
-        public System.DateTime FinishTime { get; set; }
-        public ServiceControl.Audit.SagaAudit.InitiatingMessage InitiatingMessage { get; set; }
-        public System.Collections.Generic.List<ServiceControl.Audit.SagaAudit.ResultingMessage> OutgoingMessages { get; set; }
-        public System.DateTime StartTime { get; set; }
-        public string StateAfterChange { get; set; }
-        public ServiceControl.Audit.SagaAudit.SagaStateChangeStatus Status { get; set; }
-    }
-    public enum SagaStateChangeStatus
-    {
-        New = 0,
-        Updated = 1,
-        Completed = 2,
-    }
     public class SagasController : System.Web.Http.ApiController
     {
         [System.Web.Http.HttpGetAttribute()]
@@ -362,15 +296,5 @@ namespace ServiceControl.Contracts.MessageFailures
         public MarkMessageFailureResolvedByRetry() { }
         public string[] AlternativeFailedMessageIds { get; set; }
         public string FailedMessageId { get; set; }
-    }
-}
-namespace ServiceControl.SagaAudit
-{
-    public class SagaInfo
-    {
-        public SagaInfo() { }
-        public string ChangeStatus { get; set; }
-        public System.Guid SagaId { get; set; }
-        public string SagaType { get; set; }
     }
 }
