--- src/ServiceControl.Audit/SagaAudit/SagaHistory.cs
@@ -1,18 +1,0 @@
-namespace ServiceControl.Audit.SagaAudit
-{
-    using System;
-    using System.Collections.Generic;
-
-    public class SagaHistory
-    {
-        public SagaHistory()
-        {
-            Changes = new List<SagaStateChange>();
-        }
-
-        public Guid Id { get; set; }
-        public Guid SagaId { get; set; }
-        public string SagaType { get; set; }
-        public List<SagaStateChange> Changes { get; set; }
-    }
-}
