--- src/ServiceControl.Audit/SagaAudit/SagaStateChangeStatus.cs
@@ -1,9 +1,0 @@
-namespace ServiceControl.Audit.SagaAudit
-{
-    public enum SagaStateChangeStatus
-    {
-        New,
-        Updated,
-        Completed
-    }
-}
