+++ src/ServiceControl.Infrastructure.RavenDB/Chunker.cs
@@ -1,9 +1,9 @@
-﻿namespace ServiceControl.Infrastructure.RavenDB.Expiration
+﻿namespace ServiceControl.Infrastructure.RavenDB
 {
     using System;
     using System.Threading;
 
-    static class Chunker
+    public static class Chunker
     {
         public static int ExecuteInChunks<T1, T2>(int total, Func<T1, T2, int, int, int> action, T1 t1, T2 t2, CancellationToken token)
         {
