+++ src/ServiceControl.Infrastructure.RavenDB/ServiceControl.Infrastructure.RavenDB.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net462</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="RavenDB.Database" Version="3.5.10-patch-35294" />
+    <PackageReference Include="NServiceBus" Version="7.4.2" />
+  </ItemGroup>
+</Project>
