+++ src/ServiceControl.SagaAudit/ResultingMessage.cs
