+++ src/ServiceControl.SagaAudit/SagaSnapshot.cs
