+++ src/ServiceControl.sln
@@ -86,7 +86,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceControl.Monitoring.U
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceControl.Transports.UnitTests", "ServiceControl.Transports.UnitTests\ServiceControl.Transports.UnitTests.csproj", "{45AC011D-6837-47E2-A428-E156A421E4C2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceControl.AcceptanceTesting", "ServiceControl.AcceptanceTesting\ServiceControl.AcceptanceTesting.csproj", "{134DD03F-F7EA-47D3-B7A9-D5514C5D75EE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceControl.AcceptanceTesting", "ServiceControl.AcceptanceTesting\ServiceControl.AcceptanceTesting.csproj", "{134DD03F-F7EA-47D3-B7A9-D5514C5D75EE}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServiceControl.SagaAudit", "ServiceControl.SagaAudit\ServiceControl.SagaAudit.csproj", "{25FD3DDB-B847-490C-8834-C48F828B8B6F}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceControl.Infrastructure.RavenDB", "ServiceControl.Infrastructure.RavenDB\ServiceControl.Infrastructure.RavenDB.csproj", "{6064822C-4A17-4557-A856-55DD94B96B9A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -230,6 +234,14 @@ Global
 		{134DD03F-F7EA-47D3-B7A9-D5514C5D75EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{134DD03F-F7EA-47D3-B7A9-D5514C5D75EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{134DD03F-F7EA-47D3-B7A9-D5514C5D75EE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{25FD3DDB-B847-490C-8834-C48F828B8B6F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{25FD3DDB-B847-490C-8834-C48F828B8B6F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{25FD3DDB-B847-490C-8834-C48F828B8B6F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{25FD3DDB-B847-490C-8834-C48F828B8B6F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6064822C-4A17-4557-A856-55DD94B96B9A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6064822C-4A17-4557-A856-55DD94B96B9A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6064822C-4A17-4557-A856-55DD94B96B9A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6064822C-4A17-4557-A856-55DD94B96B9A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
