--- src/ServiceControl/SagaAudit/Messages/SagaChangeOutput.cs
@@ -1,16 +1,0 @@
-namespace ServiceControl.EndpointPlugin.Messages.SagaState
-{
-    using System;
-    using NServiceBus;
-
-    class SagaChangeOutput : ICommand
-    {
-        public string MessageType { get; set; }
-        public DateTime TimeSent { get; set; }
-        public DateTime? DeliveryAt { get; set; }
-        public TimeSpan? DeliveryDelay { get; set; }
-        public string Destination { get; set; }
-        public string ResultingMessageId { get; set; }
-        public string Intent { get; set; }
-    }
-}
