--- src/ServiceControl/SagaAudit/SagaInfo.cs
@@ -1,11 +1,0 @@
-﻿namespace ServiceControl.SagaAudit
-{
-    using System;
-
-    public class SagaInfo
-    {
-        public string ChangeStatus { get; set; }
-        public string SagaType { get; set; }
-        public Guid SagaId { get; set; }
-    }
-}
